zhufu
1.地图编辑器完善
2.地图刷怪完善
3.单位炮塔等逻辑完善

+++ Remnant Afterglow/src/cfg/config_class/AnimaBullet.cs
@@ -15,8 +15,8 @@ namespace Remnant_Afterglow
         /// <summary>        
         /// 帧动画类型
         ///1 默认动画
-        ///6 子弹命中
-        ///7 子弹消失
+        ///20 子弹命中
+        ///21 子弹消失
         /// </summary>
         public int AnimaType { get; set; }
         /// <summary>        
@@ -41,15 +41,22 @@ namespace Remnant_Afterglow
         /// </summary>
         public Vector2I Size { get; set; }
         /// <summary>        
-        /// 
+        /// 播放时，偏移中心的值
+        ///（横,竖）
         /// </summary>
         public Vector2 Offset { get; set; }
         /// <summary>        
-        /// 
+        /// 各帧相对持续时间
+        ///各帧默认为1
+        ///持续时间为1的帧的显示长度是持续时间为2的帧的两倍
+        ///例子:(1,2)|(3,1)
+        ///表示第一帧相对持续时间为2
+        ///表示第三帧相对持续时间为1
         /// </summary>
-        public List<List<int>> RelativeList { get; set; }
+        public List<List<float>> RelativeList { get; set; }
         /// <summary>        
-        /// 
+        /// 设置动画播放倍数
+        ///2表示两倍播放，0.5表示半速播放
         /// </summary>
         public int FrameSpeed { get; set; }
         /// <summary>        
@@ -65,11 +72,13 @@ namespace Remnant_Afterglow
         /// </summary>
         public Texture2D Picture { get; set; }
         /// <summary>        
-        /// 
+        /// 不水平翻转
+        ///不填默认为True
         /// </summary>
         public bool FlipH { get; set; }
         /// <summary>        
-        /// 
+        /// 不垂直翻转
+        ///不填默认为True
         /// </summary>
         public bool FlipV { get; set; }
 
@@ -83,7 +92,7 @@ namespace Remnant_Afterglow
 			MaxIndex = (int)dict["MaxIndex"];
 			Size = (Vector2I)dict["Size"];
 			Offset = (Vector2)dict["Offset"];
-			RelativeList = (List<List<int>>)dict["RelativeList"];
+			RelativeList = (List<List<float>>)dict["RelativeList"];
 			FrameSpeed = (int)dict["FrameSpeed"];
 			IsAutoplay = (bool)dict["IsAutoplay"];
 			IsLoop = (bool)dict["IsLoop"];
@@ -104,7 +113,7 @@ namespace Remnant_Afterglow
 			MaxIndex = (int)dict["MaxIndex"];
 			Size = (Vector2I)dict["Size"];
 			Offset = (Vector2)dict["Offset"];
-			RelativeList = (List<List<int>>)dict["RelativeList"];
+			RelativeList = (List<List<float>>)dict["RelativeList"];
 			FrameSpeed = (int)dict["FrameSpeed"];
 			IsAutoplay = (bool)dict["IsAutoplay"];
 			IsLoop = (bool)dict["IsLoop"];
@@ -123,7 +132,7 @@ namespace Remnant_Afterglow
 			MaxIndex = (int)dict["MaxIndex"];
 			Size = (Vector2I)dict["Size"];
 			Offset = (Vector2)dict["Offset"];
-			RelativeList = (List<List<int>>)dict["RelativeList"];
+			RelativeList = (List<List<float>>)dict["RelativeList"];
 			FrameSpeed = (int)dict["FrameSpeed"];
 			IsAutoplay = (bool)dict["IsAutoplay"];
 			IsLoop = (bool)dict["IsLoop"];
