+++ Remnant Afterglow/src/cfg/config_class/AnimaExplode.cs
@@ -32,15 +32,22 @@ namespace Remnant_Afterglow
         /// </summary>
         public int MaxIndex { get; set; }
         /// <summary>        
-        /// 
+        /// 播放时，偏移中心的值
+        ///（横,竖）
         /// </summary>
         public Vector2 Offset { get; set; }
         /// <summary>        
-        /// 
+        /// 各帧相对持续时间
+        ///各帧默认为1
+        ///持续时间为1的帧的显示长度是持续时间为2的帧的两倍
+        ///例子:(1,2)|(3,1)
+        ///表示第一帧相对持续时间为2
+        ///表示第三帧相对持续时间为1
         /// </summary>
-        public List<List<int>> RelativeList { get; set; }
+        public List<List<float>> RelativeList { get; set; }
         /// <summary>        
-        /// 
+        /// 设置动画播放倍数
+        ///2表示两倍播放，0.5表示半速播放
         /// </summary>
         public int FrameSpeed { get; set; }
         /// <summary>        
@@ -61,11 +68,13 @@ namespace Remnant_Afterglow
         /// </summary>
         public Texture2D Picture { get; set; }
         /// <summary>        
-        /// 
+        /// 不水平翻转
+        ///不填默认为True
         /// </summary>
         public bool FlipH { get; set; }
         /// <summary>        
-        /// 
+        /// 不垂直翻转
+        ///不填默认为True
         /// </summary>
         public bool FlipV { get; set; }
 
@@ -78,7 +87,7 @@ namespace Remnant_Afterglow
 			LengWidth = (Vector2I)dict["LengWidth"];
 			MaxIndex = (int)dict["MaxIndex"];
 			Offset = (Vector2)dict["Offset"];
-			RelativeList = (List<List<int>>)dict["RelativeList"];
+			RelativeList = (List<List<float>>)dict["RelativeList"];
 			FrameSpeed = (int)dict["FrameSpeed"];
 			PlaySpeed = (int)dict["PlaySpeed"];
 			IsAutoplay = (bool)dict["IsAutoplay"];
@@ -99,7 +108,7 @@ namespace Remnant_Afterglow
 			LengWidth = (Vector2I)dict["LengWidth"];
 			MaxIndex = (int)dict["MaxIndex"];
 			Offset = (Vector2)dict["Offset"];
-			RelativeList = (List<List<int>>)dict["RelativeList"];
+			RelativeList = (List<List<float>>)dict["RelativeList"];
 			FrameSpeed = (int)dict["FrameSpeed"];
 			PlaySpeed = (int)dict["PlaySpeed"];
 			IsAutoplay = (bool)dict["IsAutoplay"];
@@ -118,7 +127,7 @@ namespace Remnant_Afterglow
 			LengWidth = (Vector2I)dict["LengWidth"];
 			MaxIndex = (int)dict["MaxIndex"];
 			Offset = (Vector2)dict["Offset"];
-			RelativeList = (List<List<int>>)dict["RelativeList"];
+			RelativeList = (List<List<float>>)dict["RelativeList"];
 			FrameSpeed = (int)dict["FrameSpeed"];
 			PlaySpeed = (int)dict["PlaySpeed"];
 			IsAutoplay = (bool)dict["IsAutoplay"];
